SmartAnalyst/requirements.txt
pandas==2.2.1
openpyxl==3.1.2
xlrd==2.0.1
matplotlib==3.8.4
openai==1.14.2
httpx==0.27.2
python-dotenv==1.0.1
docxtpl==0.16.7
pywin32; platform_system == "Windows"
nbformat==5.10.3
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.3
celery==5.4.0
redis==5.2.0
boto3==1.35.36
PyJWT==2.9.0
python-multipart==0.0.12
pytest==8.3.3
